Fixed issue with functions not being defined in a11y scripts Fixed issue with pages with sidebars not aligning correctly
