+++ src/scripts/a11y/plugins/constant-contact-forms.js
@@ -19,4 +19,11 @@ jQuery(document).ready(function() {
 			}, 1000);
 		}
 	});
+
+	function stripeIframeAttributes(frame){
+		$(frame).removeAttr('frameborder');
+		$(frame).removeAttr('scrolling');
+		$(frame).removeAttr('allowtransparency');
+		$(frame).removeAttr('allowfullscreen');
+	}
 });
