+++ src/scripts/a11y/plugins/events-calendar.js
@@ -66,4 +66,18 @@ jQuery(document).ready(function() {
 		}	
 	});
 	
+	function rgb2hex(rgb){
+		rgb = rgb.match(/rgb\((\d+),\s*(\d+),\s*(\d+)\)/);
+		return "#" +
+		 ("0" + parseInt(rgb[1],10).toString(16)).slice(-2) +
+		 ("0" + parseInt(rgb[2],10).toString(16)).slice(-2) +
+		 ("0" + parseInt(rgb[3],10).toString(16)).slice(-2);
+	}
+	
+	function stripeIframeAttributes(frame){
+		$(frame).removeAttr('frameborder');
+		$(frame).removeAttr('scrolling');
+		$(frame).removeAttr('allowtransparency');
+		$(frame).removeAttr('allowfullscreen');
+	}
 });
