+++ src/styles/custom/index.scss
@@ -1,5 +1,4 @@
 @import 'template';
 @import 'page-title';
 @import 'page-date';
-@import 'sidebar';
 @import 'print';
